rework variant representation
Adjust the representation of variants in the database to better accommodate big
quantities of variants in a project. Instead of storing variant indices in an array field
(with -1 representing the base variant), this change introduces the field "var_type"
with possible values "base", "change" and "addition".

Elements with var_type "base" gain an additional array field "not_in_var" holding indices of
variants in which the element has been removed or changed. Types "change" and "addition" get
an additional field "variant" referencing the given variant index.

This commit covers
* creation/deletion of variants
* CRUD of elements with variants