check project version compatability on activation
Projects are versioned with the latest pandahub release which included schema changes.

If the project had schema migrations from later pandahub versions applied, trying to activate it will log an error asking the user to update pandahub.

If the project database needs migrations in order to work with the current pandahub version, they will be
applied if activate_project() is called with apply_upgrades=True, otherwise the project is not activated and an error logged to the user.