Fix: Return value of `create_elements_in_db`
The `PandaHub.create_elements_in_db` function used to return a list of
`None`s, since it used the `update` function of a Python dict, which
returns `None` and not the updated dictionary. Fix this by using the
overloaded `|` operator for dictionaries instead, so that
`create_elements_in_db` returns a list of dictionaries with the
properties of the newly created elements.